Added option to allow customizing json serialization settings
