+++ Spinit.CosmosDb/DatabaseOptions.cs
@@ -1,4 +1,7 @@
-﻿namespace Spinit.CosmosDb
+﻿using Newtonsoft.Json;
+using System;
+
+namespace Spinit.CosmosDb
 {
     public class DatabaseOptions<TDatabase> : IDatabaseOptions
         where TDatabase : CosmosDatabase
@@ -16,5 +19,10 @@
         /// Optional preferred location
         /// </summary>
         public string PreferredLocation { get; set; }
+
+        /// <summary>
+        /// Optional Json serialization settings configuration
+        /// </summary>
+        public Action<JsonSerializerSettings> ConfigureJsonSerializerSettings { get; set; }
     }
 }
